feat: add nagiostats exporter collection method
Collecting metrics with the `nagiostats` binary isn't as dynamic as
using Nagios XI's RESTful API, but provides a way to collect metrics on
nearly any flavor of Nagios. The metric collection is significantly
harder to maintain than using Nagios XI's API, and the amount of metrics
that can be collected is limited, but the essential metrics work.