feat: add NagiosXI update metric
This adds an optional metric (`nagios_update_available_info`) to
determine whether a NagiosXI update is available. We get the version
from the NagiosXI API, so it's just a matter of the exporter reaching
out to `https://assets.nagios.com/downloads/nagiosxi/versions.php`,
fetching the latest version, and comparing it to the current version.

A metric value of `1` indicates an update is available. If the flag
`--nagios.check-updates` is not provided, then the metric value will
always be `0`. The metric collection is gated behind a flag because
users may not want their exporter reaching out to the public internet
repeatedly.